+++ R/thorn.R
@@ -4,7 +4,8 @@
 #'
 #' @param shader the name of the shader, one of \code{"thorn"},
 #'   \code{"thorn-color"}, \code{"ikeda"}, \code{"sweet"}, \code{"biomorph1"},
-#'   \code{"biomorph2"}, or \code{"biomorph3"}
+#'   \code{"biomorph2"}, \code{"biomorph3"}, \code{"apollony"}, \code{"smoke"},
+#'   \code{"plasma"}
 #' @param width,height a valid CSS measurement (like \code{"100\%"},
 #'   \code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
 #'   string and have \code{"px"} appended
@@ -16,8 +17,9 @@
 #'
 #' @examples library(thorn)
 #' thorn("ikeda") # click on the shader to animate it
+#' thorn("thorn") # you can also use the mouse wheel on this one
 #'
-#' # four shaders: ####
+#' # four shaders ####
 #' library(htmltools)
 #'
 #' hw1 <- thorn("thorn-color", width = "50vw", height = "50vh")
@@ -58,7 +60,10 @@ thorn <- function(
       "biomorph1",
       "biomorph2",
       "biomorph3",
-      "sweet"
+      "sweet",
+      "apollony",
+      "smoke",
+      "plasma"
     ))
   )
 
@@ -142,7 +147,9 @@ thorn <- function(
 #'             "biomorph1",
 #'             "biomorph2",
 #'             "biomorph3",
-#'             "sweet"
+#'             "sweet",
+#'             "apollony",
+#'             "smoke"
 #'           )
 #'         )
 #'       )