+++ NEWS.md
@@ -1,3 +1,7 @@
+# thorn 0.2.0
+
+Added three shaders: 'apollony', 'plasma', and 'smoke'.
+
 # thorn 0.1.0
 
 First release.